Create factors
Reads file line by line, then converts each into an integer. It calls a function to find the factors of the number. The function uses a loop to divide the number by increasing integers starting from 2 until the square root of the number, if the number is divisible by the integer, it is added  to the list of factors. This continues until the number is 1.